requirements.txt
beautifulsoup4
selenium
requests
pandas
numpy
nltk
scikit-learn
tensorflow
gspread
google-auth
openpyxl